refactor: change folder directory
