+++ components/tab/global/webtoon.js
@@ -1,7 +1,8 @@
-import { createCategory } from "../../category/index.js";
+import { createCategory } from "../../common/category/index.js";
 import { LIST } from "../../../js/constants.js";
 import { $ } from "../../../js/utils.js";
 import { createTabWebtoonHome } from "../sub/webtoonHome.js";
+import { SNBHandler } from "../../../js/handler.js";
 
 const category = createCategory(LIST.CATEGORIES, "snb");
 
@@ -9,5 +10,8 @@ export const tabWebtoonHome = {
   render() {
     $(".container").innerHTML = category + $(".container").innerHTML;
     $(".contents").innerHTML = createTabWebtoonHome();
+  },
+  addEvent() {
+    $(".snb").addEventListener("click", SNBHandler);
   },
 };
