+++ js/event.js
@@ -5,10 +5,6 @@ export function addEventGNB() {
   $("ul").addEventListener("click", GNBHandler);
 }
 
-export function addEventSNB() {
-  $(".snb").addEventListener("click", SNBHandler);
-}
-
 export function removeEventSNB() {
   $(".snb").removeEventListener("click", SNBHandler);
 }
